+++ src/services/agents/index.ts
@@ -1,0 +1,1 @@
+export { FeedbackService as Feedback, FeedbackService } from './feedback/feedback';
